feat: Extract useAudio
