+++ src/pages/ProjectText/ProjectVoiceSelect/hooks.ts
@@ -1,31 +1,9 @@
-import { useCallback, useEffect, useRef, useState } from 'react';
+import { useCallback, useState } from 'react';
 import {
   IVoiceSelect,
   KVoiceSelectServerValueMapper,
 } from '~/types/project/voices';
 import { findLangSexKey } from '~/pages/ProjectText/ProjectVoiceSelect/constants';
-
-export const useAudio = () => {
-  const audio = useRef<HTMLAudioElement | null>(null);
-  const play = useCallback((src: string) => {
-    audio.current?.pause();
-    audio.current = new Audio(src);
-    audio.current?.play();
-  }, []);
-
-  const stop = useCallback(() => {
-    audio.current?.pause();
-    audio.current = null;
-  }, []);
-
-  useEffect(() => {
-    return () => {
-      stop();
-    };
-  }, []);
-
-  return { play, stop };
-};
 
 export const useOptions = (
   defaultOptions: IVoiceSelect,
